Move Message classes to \Mdanter\Ecc\Message Remove static method EccFactory::getMessage()
