+++ src/EccFactory.php
@@ -2,7 +2,7 @@
 
 namespace Mdanter\Ecc;
 
-use Mdanter\Ecc\Crypto\Message;
+use Mdanter\Ecc\Message\Message;
 use Mdanter\Ecc\Crypto\Signature\Signer;
 use Mdanter\Ecc\Curves\NistCurve;
 use Mdanter\Ecc\Curves\SecgCurve;
@@ -75,16 +75,6 @@ class EccFactory
     public static function createCurve($prime, $a, $b, MathAdapterInterface $adapter = null)
     {
         return new CurveFp($prime, $a, $b, $adapter ?: self::getAdapter());
-    }
-
-    /**
-     * @param $content
-     * @param $algorithm
-     * @return Message
-     */
-    public static function getMessage($content, $algorithm, MathAdapterInterface $adapter = null)
-    {
-        return new Message($adapter ?: self::getAdapter(), $content, $algorithm);
     }
 
     /**
