+++ src/Message/MessageFactory.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Mdanter\Ecc\Crypto;
+namespace Mdanter\Ecc\Message;
 
 use Mdanter\Ecc\Math\MathAdapterInterface;
+use Mdanter\Ecc\Message\Message;
 
 class MessageFactory
 {
